+++ plugins/upload/src/lib.rs
@@ -21,7 +21,10 @@ use tauri::{
     plugin::{Builder as PluginBuilder, TauriPlugin},
     Runtime,
 };
-use tokio::{fs::File, io::AsyncWriteExt};
+use tokio::{
+    fs::File,
+    io::{AsyncWriteExt, BufWriter},
+};
 use tokio_util::codec::{BytesCodec, FramedRead};
 
 use read_progress_stream::ReadProgressStream;
@@ -74,7 +77,7 @@ async fn download(
     let response = request.send().await?;
     let total = response.content_length().unwrap_or(0);
 
-    let mut file = File::create(file_path).await?;
+    let mut file = BufWriter::new(File::create(file_path).await?);
     let mut stream = response.bytes_stream();
 
     while let Some(chunk) = stream.try_next().await? {
@@ -84,6 +87,7 @@ async fn download(
             total,
         });
     }
+    file.flush().await?;
 
     Ok(())
 }